[PORT FROM R2]WLAN: make SSID list refresh more smooth
BZ: 22930 17650

Today expiration scan count is not set.
So its default value (=2) is used.
After three scan result without the same AP,
it is removed from SSID list.

Increase expiration scan count to 10.

As scan occurs every 10 sec when in WifiSetting
or 40 sec otherwise,  usualy it will be the temporal
aging (previouly set to 1 minute) that will be used.

As a consequence, SSID list when in WifiSettings should
not be empty so quickly.

Change-Id: I964d15baebbf5b277bcaf64da8fec4dd4e2c001c
Orig-Change-Id: Id91c96a596a9cedb8128408877fb1452e0c5fca8
Signed-off-by: Honore Tricot <honorex.tricot@intel.com>
Reviewed-on: http://android.intel.com:8080/37289
Reviewed-by: Trivelly, JeanX <jeanx.trivelly@intel.com>
Reviewed-by: Garcia, JeremieX <jeremiex.garcia@intel.com>
Reviewed-by: Regairaz, JulieX <juliex.regairaz@intel.com>
Reviewed-by: Bachot, Jean-michelX <jean-michelx.bachot@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>