[PORT FROM R2] Set additional platform variables
BZ: 16882

The variables:
  TARGET_BOOTLOADER_BOARD_NAME
  TARGET_BOARD_PLATFORM
  property: ro.com.google.clientidbase=android-google

are reflected in system properties and should be set. These variables may be
changed to more appropriate values in customer builds.

Note also that we are still using 'mrst' for TARGET_BOARD_PLATFORM.
This is mostly historical, but if changed will break code that builds
shared libraries (for HAL components).

Change-Id: Icf019121c89d23fd1d1796efce9cf7789734431f
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Reviewed-on: http://android.intel.com:8080/41435
Reviewed-by: Tardy, Pierre <pierre.tardy@intel.com>
Tested-by: Letessier, ChristopheX <christophex.letessier@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>