Wifi: Inserting wl12xx and wl12xx_sdio modules on init
BZ: 20626

Currently, wifi modules (wl12xx and wl12xx_sdio) are loaded each
time the wifi is enabled. This is not recommended from the kernel
point of view since the wifi is seen as a non-removable sdio card.

So the solution is to load those modules in oneshot at boot time
before performing the provisionning step.

Change-Id: I5f2fdefc5ed0b8f2ba7976b9006c6bdc7bfad0d4
Signed-off-by: Jeremie GARCIA <jeremiex.garcia@intel.com>
Signed-off-by: Aymen Zayet <aymenx.zayet@intel.com>
Reviewed-on: http://android.intel.com:8080/42701
Reviewed-by: Trivelly, JeanX <jeanx.trivelly@intel.com>
Reviewed-by: Regairaz, JulieX <juliex.regairaz@intel.com>
Reviewed-by: Desurmont, DavidX <davidx.desurmont@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>