REVERT ME: BOOTCLASSPATH specialization by platform variant - common
BZ: 30692

This is a temporary workarround in order to specialize BOOTCLASSPATH for LEX Phone.

As BOOTCLASSPATH has to be set in the "init" stage of init script, we remove it from the commom init.rc.
Hence we set it in all init.<xxx>.rc.
This is done to work arround the "one shot" behavior of the "export" command in this scripting language.

Change-Id: I52174d5300ecd6ebf73b7974c58b862b5677ef70
Signed-off-by: Nicolas SUET <nicolas.suet@intel.com>
Reviewed-on: http://android.intel.com:8080/43270
Reviewed-by: Berthier, Emmanuel <emmanuel.berthier@intel.com>
Reviewed-by: Peix, FabienX <fabienx.peix@intel.com>
Reviewed-by: Desurmont, DavidX <davidx.desurmont@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>