VOIP on LEX: enabling of all SW supported accoustics effects by default
BZ: 31716

The purpose of this patches set is to enable all accoustics effects supported
by platform for a VOIP call.
The effects are described in audio_effects.conf file. This file contains the
list of effects (name, ID, etc.), the list of libraries which implement the
effects, etc.
By default, this configuration file is stored in /system/etc/ but a vendor
over-writes it if it stores the file in /vendor/etc; i.e. EffectsFactory of
audio framework reads first /vendor/etc/audio_effects.conf and if it does not
exist, then, it reads /system/etc/audio_effect.conf
In case of VOIP call on LEX, the effects are SW-implemented by WebRTC through
Audio PreProcessing library. In case of products as MFLD, the effects are done
by a dedicatted chip and cannot be managed through a library, i.e. because
the source code is same for all products, in case of MFLD, it is useless to
activate the SW effects, i.e. a dummy/stub Audio PreProcessing library is
necessary.

Among modifications:
 - first one defines this stub audio preprocessing library to use on MFLD
 - next one, defines for each product a specific audio_effect.conf file and
copy this file to /vendor/etc of phone FS.

Change-Id: Iaa0d884684436cfc7ce702ed9d7f0df4c3036b2b
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Reviewed-on: http://android.intel.com:8080/43982
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>