ANDROID-COS: Set USB property in COS
BZ: 49478

This patch configures USB mode as "charging" in COS.
Without configuring this, Windows detects the platform as Unknown-device,
and SDP may refuse power-supply to an unknown-device.
Also, when device is in unconfig-state, it is not allowed to draw 500mA.

"charging", is similar to "adb" value, but without starting adbd.
This is added in vendor/intel/<platform>/init.usb.rc

Change-Id: Id37d340fb3353021c1e6e8a9c2b0953d6b7cea3f
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>
Reviewed-on: http://android.intel.com:8080/59540
Reviewed-by: Beare, Bruce J <bruce.j.beare@intel.com>
Tested-by: P, Eshwar N <eshwar.n.p@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>