init.rc: increase max buffersize for EDGE
BZ: 45715

With the current edge values, we are not reaching the max
bandwidth for EDGE MCS9.

Because the value of the max buffersize is too low, this is
causing multiple releases of the TBF due to the T3182 timer.
Thus the max bandwith when using 4 timeslots uplink is not
reached.

Change-Id: I79e59eacbf47f139d7571803106fda1b1bf0e632
Origin-Change-Id: I6971b60b7f3f9022e32a4b30803d458235f10564
Signed-off-by: Fabien Hubert <fabienx.hubert@intel.com>
Reviewed-on: http://android.intel.com:8080/59977
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>