[Config]remove custo audio_effects.conf
BZ: 49721

The purpose of this patches set is to add to configuration files
the propriety which indicates if platform has or has not the
Audience chip. This propriety is necessary in case of acoustics effects
which can be done by HW chip like Audience or by SW effects (webrtc,
cristal, etc.)
Because of this propriety, the library and customized audio_effect.conf file
of BZ#31716 are not more necessary.

Change-Id: I5f3fe3bafc163a8e3b8d0ab2518bb470044e2007
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Reviewed-on: http://android.intel.com:8080/62531
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>