EM/Fuel Gauge : Save and restore fuel gauge data
BZ: 51500

Fuel Gauge needs to be configured based on the model name. Also the
learned data should be saved so that a battery removal doesn't result
in reporting wrong capacity. The config files are in user space. So
we are using an app to initialize the FG and saving the learned data

Change-Id: I26a59533614f5ea37579616a5b22c9bf80ecdec4
Signed-off-by: Jenny TC <jenny.tc@intel.com>
Reviewed-on: http://android.intel.com:8080/61990
Reviewed-by: Gopal, Saranya <saranya.gopal@intel.com>
Reviewed-by: G S, RajkumarX <rajkumarx.g.s@intel.com>
Reviewed-by: Tardy, Pierre <pierre.tardy@intel.com>
Tested-by: P, Eshwar N <eshwar.n.p@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>