[PORT FROM R3] update buffersize for gprs and edge
BZ: 48607

Systematic download of music failed on orange network

Lots of tcp duplicate ack during the download
which leads to a reset by server.

Proposal is to change buffersize to have smaller window
size for the read. Default values from android and R4 JB
are used for min values and default values (read and write)
Then max windows size are increased to reach the target
for each technology

Change-Id: Ifcae75f27359bd483fccd2043850d36f291da3c7
Orig-Change-Id: I2bcc07a2240bf3569aaac3a29f54d3b67930cc26
Signed-off-by: Fabien Hubert <fabienx.hubert@intel.com>
Signed-off-by: sami lebsir <samix.lebsir@intel.com>
Reviewed-on: http://android.intel.com:8080/65627
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>