external: hook the build metatarget for external release
BZ: 79013

This directory contains hooks for making external releases's prebuilt binaries

The makefile variables like BUILD_SHARED_LIBRARY are changed in order to be customized
and add our own Makefile targets so that we can copy the header files and binaries to the
prebuilt directory in order to be published aside in a binary artifact storage server.

The external.mk file must be included in the begining of the BoardConfig.mk files, so that
the hooks are done before any Android.mk inclusion. (actually this is factorized in common dir)

Thus we are in vendor/intel/common, and the include external/external.mk is made in common/BoardConfig.mk

when the GOAL intel_prebuilts is set, and when the "LOCAL_PATH" variable contains "/PRIVATE/",
then the metatargets are hooked, so that the proper rules are generated for creating prebuilt makefiles.

Change-Id: Ie5c72fe9bd91554affd14134fac8d36eb2c76f3f
Signed-off-by: Pierre Tardy <pierre.tardy@intel.com>
Reviewed-on: http://android.intel.com:8080/85888
Reviewed-by: Monnier, Olivier <olivier.monnier@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>