[REVERT ME]Temporary runtime_pm dynamic disabling for Bluetooth
BZ: 81978

Changing permissions to allow libbt-vendor to write to the
power control interface for dynamically disabling the
runtime_pm. It's part of a temporary hack to have Bluetooth
working on boards that have a wrong hardware configuration.

This change has to be reverted as soon as the correct
HW configuration is available.

Change-Id: I3533495a3062d64152fe6d798e000c6372b2bf04
Signed-off-by: Raffaele Aquilone <raquiloX>
Reviewed-on: http://android.intel.com:8080/87904
Reviewed-by: Aquilone, RaffaeleX <raffaelex.aquilone@intel.com>
Reviewed-by: Gros, Julien <julien.gros@intel.com>
Reviewed-by: Seibel, Eric <eric.seibel@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>