EM/FG_CONFIG: Prevent FG characterization data write to chip in COS
BZ: 79737

The issue is when in COS, we see the config_init is 0, read from the
primary config file. Due to this the FG driver writes the characterization
data to the chip. To prevent this to happen, config_init is set to 1
in COS mode, which prevents writing the characterization data to chip.

Change-Id: Ib2f54960c32c18fee6f1bcddada48d52f20c6960
Signed-off-by: R, Kannappan <r.kannappan@intel.com>
Reviewed-on: http://android.intel.com:8080/89809
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Reviewed-by: Thakur, GauravX <gauravx.thakur@intel.com>
Reviewed-by: Kumar, AshutoshX <ashutoshx.kumar@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>