Changed a sdcard parameter from size_hint to #size_hint.
BZ: 89060

(patch re-submitted as a first re-integration step to avoid FOTA issue)
That change was made to generate a recovery.fstab that is standard,
and not an Intel customized version to support MBR partitioning.
With that change, the size_hint is specified as a comment, which
makes it still usable for Intel, and ignored by default Android code.

Change-Id: I87628902428d3888aa75f66da543c1333549e194
Signed-off-by: Florent Auger <florent.auger@intel.com>
Reviewed-on: http://android.intel.com:8080/97115
Reviewed-by: Jauffres, MichelX <michelx.jauffres@intel.com>
Reviewed-by: Pince, Jean-christophe <jean-christophe.pince@intel.com>
Tested-by: Molero Rodenas, BeatrizX <beatrizx.molero.rodenas@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>