Restore COLD_BOOT boot type for ROS to MOS transition
BZ: 89137

When executing adb reboot to go from ROS to MOS a COLD_RESET
was done instead of a COLD_BOOT.
COLD_BOOT should have also been done from droidboot to MOS.

Error is due to a bad passed value to force_cold_boot
parameter.
In R4 main it must be 1 (FORCE COLD_BOOT) or 0 (NO COLD_BOOT)
instead of the Y or N that is given for R4.1.

Change-Id: I9cf241ab221bf8fa45c0b4dce1e177ae79311569
Signed-off-by: Michel Jauffres <michelx.jauffres@intel.com>
Reviewed-on: http://android.intel.com:8080/95290
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Afantenos, Marie Cecile <marie.cecile.afantenos@intel.com>
Reviewed-by: Castelain, DavidX <davidx.castelain@intel.com>
Reviewed-by: Imberton, Guilhem <guilhem.imberton@intel.com>
Tested-by: Afantenos, Marie Cecile <marie.cecile.afantenos@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>