init: Remove setprop system_init.startsurfaceflinger 1
BZ: 91727

Needed by:
    vendor/intel/baytrail - Ia1b58b377e2fab07fdbf498d777c362f2c4cb211

For some targets it is desirable to run surfaceflinger as an internal
thread within system_server, whereas for other targets it is desirable
to run surfaceflinger as a separate process.

If property system_init.startsurfaceflinger is not defined, or if
it is defined as "1", then system_server will start its internal
surfaceflinger thread.  Otherwise, not.

Previously, the common version of init.rc unconditionally set property
system_init.startsurfaceflinger to "1".  This was not only unnecessary
(as that is the default anyway), but also made setting a different
value in a board-specific init.xxx.rc difficult.

Therefore, remove:
    setprop system_init.startsurfaceflinger 1

A board-specific init file may cause surfaceflinger to be run as a
separate process by:
- setprop system_init.startsurfaceflinger 0
- Defining service surfaceflinger.

Change-Id: I03edbc9b54b150afc24b2426b5e675887bb50f6f
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>
Reviewed-on: http://android.intel.com:8080/96477
Reviewed-by: Luc, Piotr <Piotr.Luc@intel.com>
Reviewed-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Tested-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>