WiFi: Remove netd from wifi_bcm.
BZ: 97757

By default, netd is built. It is listed in
build/target/product/base.mk.
So rename netd.bcm to netd in order to compile it
by default. It is required to have it in the board
to be able to boot. Even if no WiFi module is selected.
So remove netd.bcm from wifi_bcm target.

Change-Id: Id8d2f00423288bdbfac5214433a08ea3d094c7b5
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Reviewed-on: http://android.intel.com:8080/100744
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Fert, Laurent <laurent.fert@intel.com>
Reviewed-by: Regairaz, Julie <julie.regairaz@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>