wifi: remove buggy dynamic chip autodetection
BZ: 99486

Curent dynamic chip autodetection relies on a successive
insmod/ifup of all BCM drivers embedded on target.
However, we see ransdom issues doing that when we try to
insmod bcm4335 after bcm4334.
As this method is proven as not reliable, we get back to
a static solution waiting for a "real" chip detection to
be implemented.

Change-Id: I1b1d78a44dabf2fd2687b8fadfc4b27718b7f502
Signed-off-by: Jeremie GARCIA <jeremiex.garcia@intel.com>
Reviewed-on: http://android.intel.com:8080/101056
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Brouat, Claude <claude.brouat@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>