[AUDIO] Add PSI Recorder APK by default on eng and user-debug builds for all projects
BZ: 99797

The PSI Recorder APK must be installed by default on eng and user-debug builds for all projects.
This patch aims to update the way the PSI Recorder APK is added in projects.
Today, the PSI Recoder APK product package is defined in each board Makefile.

Now, we will define the PSI Recoder APK product package in the common Makefile.

Change-Id: I2d0de6324a8b7f98fb05f9628e7927a09bd709a2
Signed-off-by: Neil Di Folco <neil.di.folco@intel.com>
Reviewed-on: http://android.intel.com:8080/101271
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Afonso, PhilippeX <philippex.afonso@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: Fert, Laurent <laurent.fert@intel.com>
Reviewed-by: Fourdan, Olivier <olivier.fourdan@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>