[WM8958] Property ro.AudioComms.PFW.ConfPath should not be read only
BZ: 103991

The Android property ro.AudioComms.PFW.ConfPath is used to give the
pathname of the main PFW configuration file. It should be possible to
set the correct path depending on runtime parameters, thus this
property should not be readonly.

Changed the name of the property to AudioComms.PFW.ConfPath

Change-Id: I734808ff1672167f72a6a258dedad29a3b081090
Signed-off-by: Frederic Boisnard <fredericx.boisnard@intel.com>
Reviewed-on: http://android.intel.com:8080/110282
Reviewed-by: Betous, GuillaumeX <guillaumex.betous@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>