[AUDIO] Remove SPEECH_RX port management from the parameter framework
BZ: 110425

When in call, if for any reason the parameter-framework is restarted
(e.g media server restart), applying default configurations from the
blackboard will lead to disabling the SPEECH_RX port.
This will result in permanant audio loss in downlink during the active call.

As per IMC recommendations, SPEECH_RX port enabling should not be done
outside of the modem itself. So this patch's purpose is to definitely
remove the parameter from the IMC subsystem.

Change-Id: Ibf0722962d1f2b09c5aab6069ecf8d5c16736678
Signed-off-by: pdiopX <piotrx.diop@intel.com>
Reviewed-on: http://android.intel.com:8080/109253
Reviewed-by: Denneulin, Guillaume <guillaume.denneulin@intel.com>
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Rocard, KevinX <kevinx.rocard@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>