Change permission on SW Sync framework
BZ: 114365

Enable access to the synchronization framework by changing /dev/sw_sync
mode to 0660 and owner to system:system.

The sw sync framework is google's example implementation of an
android sync framework driver. As such it shouldn't be used in
production systems, however for testing purposes this set of
changes is being uploaded so other developers can explore the
framework.

Change-Id: I1ad431b0c62393aa1eb2687b09c81a9e6a01e9d0
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>
Signed-off-by: Fabien Marotte <fabien.marotte@intel.com>
Reviewed-on: http://android.intel.com:8080/114275
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>