Change permission on SW Sync framework
BZ: 114365

Enable access to the synchronization framework by changing /dev/sw_sync
mode to 0660 and owner to system:system.

The sw sync framework is google's example implementation of an
android sync framework driver. As such it shouldn't be used in
production systems, however for testing purposes this set of
changes is being uploaded so other developers can explore the
framework.

Change-Id: I40a8009929f8c18348bf6a7918473147afcb2cf9
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>
Reviewed-on: http://android.intel.com:8080/111457
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Smith, Gary K <gary.k.smith@intel.com>
Reviewed-by: Fert, Laurent <laurent.fert@intel.com>
Reviewed-by: Barbalho, Rafael <rafael.barbalho@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>