PowerSubsystem XML Structure
BZ: 119640

Today, during voice call, merrifield board will try to enter
suspend. This is not permitted as accoustics are run on LPE
and s0i2-LPE is the shallowest state allowed then.

This patch introduces the PowerSubsystem plugin, allowing the PFW
to manage wakelocks in order to prevent the system to enter into
suspend mode when some conditions are met.

Change-Id: Ida3a6ecce4fc305885827da5c6763f3fb48738c2
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>
Reviewed-on: http://android.intel.com:8080/117309
Reviewed-by: Rocard, KevinX <kevinx.rocard@intel.com>
Reviewed-by: Denneulin, Guillaume <guillaume.denneulin@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>