PM: Changing the kernel commandline for user build
BZ: 107003

The watchdogtimer is causing wakeups in s0ix state.This is
leading to an increase in s0ix power which needs to be optimized.
In victoriabay for instance, it is giving a power savings of 5-6mW.
Hence, we are changing the watchdog timer threshold for
hardlockups in the command line for user builds alone as the lockup
detectors are essentially a debug feature & 60 secs is the max value
of the threshold.

Signed-off-by: Ashish K <ashish.k@intel.com>
Change-Id: I0cb9a820f0a220c4edb4cd09c74cfcf0aff2bf3f
Reviewed-on: http://android.intel.com:8080/118108
Reviewed-by: Mansoor, Illyas <illyas.mansoor@intel.com>
Tested-by: Delance, OlivierX <olivierx.delance@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>