[Audio] PFW: Prevent init.common.rc from updating the AudioComms.PFW.ConfPath property
BZ: 126047

During MTBF tests on Saltbay, an issue has been raised about PFW
startup. Sometimes, the main XML file used for the PFW is not the
correct one.
The following file is used:
"/etc/parameter-framework/ParameterFrameworkConfiguration.xml"
... instead of:
"/etc/parameter-framework/ParameterFrameworkConfiguration-boomer.xml".

The PFW is started by the RouteManager which retrieves this XML
file path thanks to the Android property "AudioComms.PFW.ConfPath":
- The default value of this property is set in the file "init.common.rc",
  under the "on boot" section,
- This default value is also defined in the RouteManager, in case the
  property is not found,
- On Saltbay, the value is updated by the HAL (before the PFW starts),
  to add the "-boomer" or "-wm8958" suffix.

In the logs, we can see that the property is correctly set by the
HAL, but right after it is reset to its default value by another process.
The only candidate for this update is the "init.common.rc" file, which
suggests a race condition between the 2 mechanisms.

This patch fixes the issue by removing the property update from the
"init.common.rc" file:
- On Saltbay, the value will now only be updated by the HAL,
- On other platforms, the RouteManager will fallback to the default
  value of the property ("/etc/parameter-framework/ParameterFrameworkConfiguration.xml")

Change-Id: I6ef38a894b5a8b28f026403a303e0127080510bc
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>