SARManager: dynamically enable/disable RF cutback notifications
BZ: 128434

Design for this patch set is:
Add a "notifications" field in sarconfig.xml so that it can be
changed dynamically.
Field to be added is:
<output name="cell" class="com.intel.internal.telephony.
SARManager.CellOutput" notifications="false"></output>
or:
<output name="cell" class="com.intel.internal.telephony.
SARManager.CellOutput" notifications="true"></output>
The way it is added in sarconfig.xml aims at having a
notification setting section for the output items.

Change-Id: I3026f585c4fddbbd2ffad641ed01ce0e285d9076
Signed-off-by: William Begoc <williamx.begoc@intel.com>
Reviewed-on: http://android.intel.com:8080/125041
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Seibel, Eric <eric.seibel@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>