fstab: factory mounted partition
BZ: 121951

Sometimes the factory partition gets corrupted and we have no solid
clue yet of what happened. The problem is, when the factory partition
is corrupted, the Filesystem Manager stops mounting the other
partitions, thus the system partition is not mounted and the Android
system is not able to start correctly and kind of freeze from the user
point of view. In order to guarantee the device is usable anyway and
since the factory partition contains only meaningless data (no data at
all actually) on FFRD10 devices, this patch makes the factory
partition to be mounted at last.

Change-Id: Ieb82581ac8c233b782f36cd147f4ffbab6c2988e
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: http://android.intel.com:8080/129007
Reviewed-by: Chouleur, Sylvain <sylvain.chouleur@intel.com>
Reviewed-by: Tinelli, Vincent <vincent.tinelli@intel.com>
Reviewed-by: Fert, Laurent <laurent.fert@intel.com>
Tested-by: Lumbierres, PhilippeX <philippex.lumbierres@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>