fstab: factory mounted partition
BZ: 181760

Currently the factory partition is mounted at first.  If the mount
manager fails to mount it, it stops mounting any partition which
prevent the other partition to be mount.  This patch changes the
partition mount order of the factory partition hence it will not
prevent the other partitions to be mounted anymore.

Change-Id: I29e0306882caef076132beebc82106c025bfd02f
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>