[WIFI] increase 'tcp_limit_output_bytes' to fix TCP upload speed
BZ: 144225

The TCP upload speed decreased from about 180Mbps to 110Mbps
because of the introduction of TCP Small Queues (TSQ) in
kernel 3.6. TSQ was introduced to combat buffer bloat, see
http://lwn.net/Articles/506237/ for details. However the default
transmit buffers size (128K) is too small for WiFi, hence increase
the size to 384K.
At the same time, limit the 'tcp_wmem' sizes to the same amount, in
order to not stop/start transmit queues.

Change-Id: I6c438b34f8214591c125019da371fe552e0e50cc
Signed-off-by: Frode Isaksen <frodex.isaksen@intel.com>
Reviewed-on: http://android.intel.com:8080/137611
Reviewed-by: Fiat, Christophe <christophe.fiat@intel.com>
Reviewed-by: Regairaz, Julie <julie.regairaz@intel.com>
Reviewed-by: Seibel, Eric <eric.seibel@intel.com>
Reviewed-by: Casasnovas, QuentinX <quentinx.casasnovas@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>