[AUDIO] Print a message when the command parameter is used on user-debug builds
BZ: 144584

By default the connection to the PFW is disabled for user-debug builds.
However, the parameter command is still available; so using this command
on an user-debug build returns an error message:

CConnectionSocket::connect::connect: Connection refused
Unable to connnect to target :-(

This message is confusing and can lead the user to think that the PFW is
not working correctly.

This patch updates the "parameter" script for user-debug builds, so that
if the connection to the PFW is impossible the following message is displayed:
"The audio PFW remote-processor is not running (disabled by default on user-debug builds)"

Change-Id: I092875238b6238ea70a00fdc0b7277ade4781625
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>