wifi: Manage external build for BCM Kernel module
BZ: 150103

The source version of the BCM drivers are not available
in the external release.

Check for availability of the sources and point toward the
prebuilt module path if the sources are not available.

Add a convenient function to translate paths between the source
and the prebuilt versions of the module.

Category: device enablement
Domain: CWS.WIFI-Bluewoods
Origin: internal
Upstream-Candidate: no, proprietary
Change-Id: I3206010dfbc586418e0d1322f11e82092de00c1b
Signed-off-by: Olivier Fourdan <olivier.fourdan@intel.com>