init.nodebug.rc: on ramconsole, reboot
BZ: 154929

init.nodebug.rc is used as init.debug.rc on user builds. 'on
ramconsole' should not be triggered on user builds as this mode is
requested by firmware, through property 'ro.boot.mode'
(androidboot.mode in cmdline), only for debug purpose, not enabled in
user builds.

In case a debug firmware is used with an Android user build, if OS
boot to ramconsole mode, this change will trigger a reboot in main OS.

Change-Id: Ided542fa3c18397a9eb972e44bc055e618641030
Signed-off-by: Jeremy Rocher <jeremy.rocher@intel.com>