Don't mount any partition in ROS except /cache and /tmp
BZ: 153600

With KK release, there is a a new verification done when installing
a new package. It checks that there are only /cache and
/tmp partitions mounted. If not, the update fails.

This patch removes any references to other partitions except
/cache and /tmp in recovery.init.rc.
To be still able to retrieve the aplogs in ROS, this patch also
modifies the service apk_logfs. It is now launched right before
leaving ROS. And to be sure the logcat command is over before
rebooting, we use a script to launch the logcat then set another
propriety that will launch the reboot.

Change-Id: I3a4859203b9c111273d55d07b9382f29e4d16f61
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>