[Dolby DS1] Dolby app crashing recently in Kitkat
BZ: 156412

The dolby jar is not included properly in BOOTCLASSPATH
in recent builds of Kitkat due to which Dolby app is crashing
and DS1 effect is not getting applied when DS1 is enabled.
Changes made to AndroidBoard.mk to include the dolby jar.

Change-Id: I35add1998c15bd62d73a5fc1fc15e3f8af4f1d07
Category: feature differentiation
Domain: AUDIO_MEDIA-Dolby
Origin: 3rd-party Dolby
Upstream-Candidate: no, proprietary
Signed-off-by: charitardha <satya.charitardha.jayanti@intel.com>