intel_prebuilts: Export classes.jar with javalib.jar
BZ: 159921

The javalib.jar containing classes.dex (dalvik format) is currently
saved when a ("dynamic") java library is built. Such file can be used as
a standalone -to launch a service for example- but it cannot be used to
compile another java module requiring the prebuilt library.

For javac to be able to build a module depending on a prebuilt java
library, classes.jar containing all the path/to/TheClass.class files
must also be saved.

Change-Id: Ib8025664114830887063d27a8d1ade71bdc26cf3
Signed-off-by: Laurent FERT <laurent.fert@intel.com>