Improve properties and mount options when exiting ROS
BZ: 162908

Goal is to be able to mount /logs and relaunch apk_logfs service,
on sys.powerctl property set, during ROS mode, before exiting from
ROS mode.

This patch improves the process because it is been observed that
when in ROS, the "adb reboot bootloader" or "adb reboot recovery"
commands did not complete from host PC.
Patch fixes this case and also ensures that the recovery logs are
not lost when rebooting.

Change-Id: I420c99a44cecc77da7e57256de2846c724df5e24
Signed-off-by: Michel Jauffres <michelx.jauffres@intel.com>