Enable ADB Secure feature for user variants
By setting ro.adb.secure to 1, we enable the secure mode in adbd. Thus,
any connecting client will be asked to provide it's private or public
key and the user handling the device will be asked to authorize the
connecting client.
Secure ADB can be used over USB or IP (WiFi and Ethernet).

Change-Id: Iafd9b73ab99329879ddc3d739cba7a75a9a816cf
For: APDEV-896
Signed-off-by: Robert Chiras <robert.chiras@intel.com>