move DB key here
The build has been using the Google testkey in
build/target/product/security to sign the bootloader and boot
images. Put a copy here so that if we want to update to a different
key pair for DB, it is easy to do so.

Update the kernel module signing location.

Issue: ADEV-520
Change-Id: Ie1fd803d3ad01268765df2c55dd2876001475895
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/160709
Reviewed-by: Jovanovic, Radivoje <radivoje.jovanovic@intel.com>
Reviewed-by: Leung, Daniel <daniel.leung@intel.com>
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>