[Audio] HAC configuration is not applied
BZ: 172494

When using an HAC device, the associated configuration is not
applied. Instead, that is the default earpiece configuration that is
selected.

Problem comes from the fact that HacSelected criteria event is not
propagated from HAL XML to PFW.

This patch fixes the declaration of this platform state change criteria
between HAL XML and PFW.

Change-Id: I29a9a45ba0786862fd7275f4ea4bf29050d46bf2
Signed-off-by: Thierry Quervel <thierry.quervel@intel.com>