publish images:Fix to copy fastboot.img, system.img to pub folder
All the images required for fastboot flashing are published
under pub/fastboot_images folder. The images fastboot.img,
system.img are missed in the pub folder because
TARGET_USE_USERFASTBOOT is not correctly read. This patch
reads the variable correctly and copies the fastboot.img,
system.img images to the pub folder.

Change-Id: Idb986e04e30d205f97c586ce1e270c88e0c1f0c6
Signed-off-by: Ravindranath D <ravindranathx.doddi@intel.com>