Audio: Handle voice processing lock during band type handover
BZ: 181382

In order for the SST FW to manage the switch from one band to another
(eg: WB -> NB...), the FW needs a mechanism to know when all the new
parameters are sent in order to manage correctly the power-on of the
speech loops.

This patch provides parameters to deal with the voice processing
lock alsa controls. It also provides routing to toggle them at
disable/enable stages in voice call.

Change-Id: I8c673236e6b9c9fb6b648040d6bcdc9f2aab0916
Signed-off-by: Guillaume Denneulin <guillaume.denneulin@intel.com>