[AUDIO] Set input devices depending on VoIP talk configuration
BZ: 180582

Currently main mic is always selected when making a VoIP call,
resulting in a wrong configuration in the SSP timeslots.

Main and back mic are now used as criteria for setting the SSP
slots instead of output devices. They have then to be strictly
set according to the mode. For IHF and headphones devices, back
mic shall be enabled. Main mic shall only be enabled in handset
mode.

Change-Id: I9eb4d8ee8f5ebde91a64b615672475e60c28c086
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>