Audio/Scalpe: hf_sns2 pipe: handle Ambient value for SelectedInputDevice in XML HAL
BZ: 182195

The criterion SelectedInputDevice managed by the XML HAL does never set the value
Ambient. However, this value could be helpful when dealing with noise reduction
features in audio tuning.

This patch updates the configuration of the XML HAL so that this Ambient value is
set up for the SelectedInputDevice criterion. For now, it is activated when the
phone is in call with Earpiece or IHF or Headphones activated.

Change-Id: I98c8a49412a1454234d5e76997437ebf484ce467
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>