Remove DPTF startup.dv Configuration File
BMDA 481
android.security.cts.ListeningPortsTest
 - testNoRemotelyAccessibleListeningTcpPorts fail

The startup.dv file starts the DPTF web server that enables
OEM for DPTF tuning. However, this function would cause the
CTS to fail. Turning if off for official releases and we will
give OEMs out-of-band instruction on how to enable the
DPTF web server.

Change-Id: Ia0038a2a2efa2b8399c21b37214c88fd1aa6d5e2
Signed-off-by: Brian Bian <brian.bian@intel.com>
Reviewed-on: https://android.intel.com/178848
Reviewed-by: Hiremath, Vijay P <vijay.p.hiremath@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>
Tested-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>