DPTF: Fix the DPTF application loading issue
Issue: BMDA 981, BMDA 990

This fix addresses the issue that the DPTF app is not started when
the startup.dv file is not present. The file startup.dv is removed
in last build in order for us to pass CTS, however, a code change
is required in ESIF_UF to start DPTF automatically when running
in daemon mode, even without the presence of startup.dv.

Change-Id: I1263b074eb02d2d2b77227dc9a602a7bd166ec13
Signed-off-by: Brian Bian <brian.bian@intel.com>
Reviewed-on: https://android.intel.com/180919
Tested-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Deuskar, Prafulla <prafulla.deuskar@intel.com>
Reviewed-by: Sripathi, Srinivas <srinivas.sripathi@intel.com>
Reviewed-by: Hiremath, Vijay P <vijay.p.hiremath@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>