DPTF:Fix device encrytion for the pid file creatd by dptf daemon in /data
Issue:BMDA-989
Tablet encryption fails

DPTF daemon esif_ufd opens a PID lock file.
This is to ensure that multiple esif_ufd cannot run at the same time .
During encrytion when all the services are killed and /data unmounted 
vold was unable to unmount /data as the esif_ufd is still runiing 
and holding the lock file in /data/misc/dptf.

This patch creates the pid file in /mnt instead of /data

Change-Id: I517a08a35d1d1dc918d74e54e8eebb29bb9d02b1
Signed-off-by: sahashau <shaunak.saha@intel.com>
Reviewed-on: https://android.intel.com/181244
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>
Tested-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>