Security:Android KK must support secure adb as CDD
Issue: BMDA-1085

By setting ro.adb.secure to 1, we enable the secure mode in adbd. Thus,
any connecting client will be asked to provide it's private or public
key and the user handling the device will be asked to authorize the
connecting client.
Secure ADB can be used over USB or IP (WiFi and Ethernet).

Change-Id: I65cf906e6b01041bb6d8d755ee4734df41a08961
Signed-off-by: sahashau <shaunak.saha@intel.com>
Reviewed-on: https://android.intel.com/189166
Reviewed-by: Sripathi, Srinivas <srinivas.sripathi@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>
Tested-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>