Audio: Update DRC parameters of WM8958 codec
BZ: 195203

Currently, the DRC parameters is not easy to configure, since is
is simply an array of 10 integers.

This patch reworks the DRC section in order to provide named
controls allowing fine tuning of this DRC.

Note: due to an endianess issue currently beeing investigated, some
of the DRC parameters still have to be handled as integers.

Change-Id: Ifa1250009e1adb054653d7ed0d243f37c207f8c3
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>