Enhance VoIP Band Type criterion management
BZ: 191183

The Band Type for VoIP is only set upon detection of the voice
communication input source. However, some VoIP application may
not make use of this input source. It prevents from configuring
correctly LPE pre processing according to the band.

This patch manages the VoIP band in a more generic way, checking
the rate of the active input stream.

Change-Id: I72576efd1bb8e3691afcf8cdca5f360c1237e34c
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>