Do not change mode in csvcall when two output devices are selected
BZ: 204604

Modem related routes were interrupted during a MO call when we are
capturing a picture with Google Camera app.

There is no need to change mode when two output devices are selected.
So when we are in csv call and we have more than one output selected
the MODES variable will not be changed from IN_CALL to NORMAL.
By removing this treatment we will decrease the interrupted time in
the distant phone.

Change-Id: Iec4950dfc5157b088cf46e7a6fe27ce416fc3836
Orig-Change-Id: Ie73f3a74d5fe3df30efe97667725354b11baa886
Signed-off-by: Amine Kristou <aminex.kristou@intel.com>