Make MAMGR specialization selection thru Makefile
BZ: 204542

The MAMGR is specialized for modem family. Until now, it was
configured using Android properties.

The configuration of MAMGR is now done using XML files which have
to be copied on the target.

This patch introduces a new Makefile to provide common modules
to be used on product on which MAMGR is used, handling XML
configuration files management.

Change-Id: I9666037de61daa5caddfd2e37b43bee19b3435e1
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: https://android.intel.com/212457
Reviewed-by: Cahuzac, ThomasX <thomasx.cahuzac@intel.com>
Reviewed-by: Seketeli, ApeleteX <apeletex.seketeli@intel.com>