Audio: use the FBA lock while changing the BWX
There is some issue when activating/deactivating the BWX.
Indeed the bwx_speech_path parameter in SBA_VB_SET_SPEECH_PATH
command and the switch parameter in FBA_VB_BWX command are not
always correlated.

This is due to the FBA lock that is not used. The BWX is activated through
a setParameter and this patch is putting the BWX criterion
as Route Criterion. So the BWX activation will use the 5 routing steps and
the FBA lock.

Change-Id: I45e49a075de755381626e4fa6d2b59337263f11b
Depends-Change-Id: If82b76a7cb537dd7ee72d954f3775432fad3211d
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-11581
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Signed-off-by: MiguelX Gaio <miguelx.gaio@intel.com>
Reviewed-on: https://android.intel.com:443/281348
Reviewed-on: https://android.intel.com/312514
Reviewed-by: Laude, Jean-patrice <jean-patrice.laude@intel.com>
Tested-by: Laude, Jean-patrice <jean-patrice.laude@intel.com>