[WiDi Bundle] Set WiFi Display P2P go_intent to 14.
In order to be able to improve Performencies,
we need to be able to choose the P2P channel
used for WiFi Display connection. Only the
P2P Group Owner can choose it, so force the
go_intent to 14.

Change-Id: Ia7618b54dd18b1ec93ba900e087a2a57b75dffd7
Category: aosp improvement
Domain: CWS.WIFI-Perf
Origin: internal
Upstream-Candidate: no, not-aosp
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-2010
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Reviewed-on: https://android.intel.com:443/260704