Generate intel prebuilts post-build
Binaries are first generated with a full build.
Then, released binaries are packaged based on available built files.

Get jar from common folder to make sure dex is not stripped.
Check for availability of common jar rather than built module as the
dependencies are set on common jar and therefore built module is usually
not generated.

Change-Id: Id578347f60eea251fec6a8eb5e3942aa08586b96
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-3831
Signed-off-by: Laurent FERT <laurent.fert@intel.com>
Reviewed-on: https://android.intel.com:443/282807